feat(io): move biblatex and i/o to dedicated crate
Move BibLaTeX mapping and RIS/JSON/YAML loading logic from citum-engine
to a new citum-io crate. Decouple citum-engine from biblatex and
ciborium dependencies to maintain format agnosticism.

Update CLI and tests to use the new crate. Fix documentation warnings,
resolve review feedback on JSON set parsing, and archive the completed
bean.

Refs: csl26-xio1, #651
