+++ crates/citum-engine/Cargo.toml
@@ -16,10 +16,8 @@ csl_legacy = { package = "csl-legacy", path = "../csl-legacy" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 serde_yaml = { package = "serde_yaml_ng", version = "0.10" }
-ciborium = "0.2"
 thiserror = "2.0"
 citum_edtf = { package = "citum-edtf", path = "../citum-edtf", features = ["serde"] }
-biblatex = "0.11"
 url = { version = "2.5", features = ["serde"] }
 indexmap = { version = "2.2.6", features = ["serde"] }
 regex = "1.10"
@@ -36,6 +34,7 @@ ffi = []
 [dev-dependencies]
 criterion = { version = "0.5", features = ["html_reports"] }
 citum_migrate = { package = "citum-migrate", path = "../citum-migrate" }
+citum_io = { package = "citum-io", path = "../citum-io" }
 roxmltree = "0.20"
 rstest = "0.25"
 tempfile = "3.8"
